backend/requirements.txt
django
firebase-admin
djangorestframework
python-decouple==3.3
django-cors-headers